@@ -1,4 +1,3 @@
-
 import 'package:tetris/ai/ai.dart';
 import 'package:tetris/ai/phenotype.dart';
 import 'package:tetris/model/game.dart';
@@ -9,51 +8,54 @@ const kStatsLinesCompleted = 'linesCompleted';
 const kStatsLastTetroHeight = 'tetrominoHeight';
 
 class Smart extends Pajitnov with MovePlayer {
-
   // https://pdfs.semanticscholar.org/b0fe/1ed14404db2eb1db6a777961440723d6e06f.pdf?_ga=2.183710566.1551116755.1589937221-487702206.1589937221
   static const List kFeatures = [
-    kStatsMaxHeight,          // 1
-    kStatsTotalHoles,         // 2
-    kStatsConnectedHoles,     // 3
-    kStatsLinesCompleted,     // 4
-    kStatsTotalHeightDiff,    // 5
-    kStatsMaxWell,            // 6
-    kStatsSumWells,           // 7
-    kStatsLastTetroHeight,    // 8
-    kStatsTotalBlocks,        // 9
-    kStatsWeightedBlocks,     // 10
+    kStatsMaxHeight, // 1
+    kStatsTotalHoles, // 2
+    kStatsConnectedHoles, // 3
+    kStatsLinesCompleted, // 4
+    kStatsTotalHeightDiff, // 5
+    kStatsMaxWell, // 6
+    kStatsSumWells, // 7
+    kStatsLastTetroHeight, // 8
+    kStatsTotalBlocks, // 9
+    kStatsWeightedBlocks, // 10
     kStatsWeightedHoles,
   ];
 
   final TetrisPhenotype phenotype;
   Smart({
-    this.phenotype,
+    required this.phenotype,
   });
 
   @override
-  Move selectMove(Game game, List<Move> moves) {
-
+  Move? selectMove(Game game, List<Move> moves) {
     // init
-    Move bestMove;
-    double bestScore;
+    Move? bestMove;
+    double? bestScore;
 
     // test each move
     //moves.shuffle();
     for (Move move in moves) {
-
       // play
       int initialLines = game.linesCompleted;
-      Tetromino currentTetromino = game.currentTetromino;
-      Game result = playMove(game, move, true, null);
-      Stats stats = Stats.from(game: result);
+      Tetromino? currentTetromino = game.currentTetromino;
+      Game? result = playMove(game, move, true, null);
+
+      // check
+      if (result == null || currentTetromino == null) {
+        continue;
+      }
 
       // custom stats
-      stats.setValue(kStatsLinesCompleted, result.linesCompleted - initialLines);
+      Stats stats = Stats.from(game: result);
+      stats.setValue(
+          kStatsLinesCompleted, result.linesCompleted - initialLines);
       stats.setValue(kStatsLastTetroHeight, currentTetromino.y);
 
       // calc score
       double score = 0;
-      for (int i=0; i<kFeatures.length; i++) {
+      for (int i = 0; i < kFeatures.length; i++) {
         score += phenotype.genes[i] * stats.getValue(kFeatures[i]);
       }
 
@@ -62,11 +64,9 @@ class Smart extends Pajitnov with MovePlayer {
         bestMove = move;
         bestScore = score;
       }
-
     }
 
     // done
     return bestMove;
   }
-
 }
