@@ -3,7 +3,7 @@ import 'package:tetris/model/game.dart';
 
 class DumbAI extends Pajitnov with MovePlayer {
   @override
-  Move selectMove(Game game, List<Move> moves) {
+  Move? selectMove(Game game, List<Move> moves) {
     moves.shuffle();
     return moves[0];
   }
