@@ -1,14 +1,13 @@
-
 import 'package:tetris/model/tetromino.dart';
 
 class TetrominoI extends Tetromino {
-
   @override
   TetrominoType get type => TetrominoType.i;
 
   @override
   int get height {
-    if (this.rotation == Rotation.Normal || this.rotation == Rotation.SixOClock) {
+    if (this.rotation == Rotation.Normal ||
+        this.rotation == Rotation.SixOClock) {
       return 1;
     } else {
       return 4;
@@ -21,22 +20,30 @@ class TetrominoI extends Tetromino {
   }
 
   @override
-  List<List<TetrominoType>> get blocks {
-    if (this.rotation == Rotation.Normal || this.rotation == Rotation.SixOClock) {
-      return [ [type,type,type,type] ];
+  List<List<TetrominoType?>> get blocks {
+    if (this.rotation == Rotation.Normal ||
+        this.rotation == Rotation.SixOClock) {
+      return [
+        [type, type, type, type]
+      ];
     } else {
-      return [ [type], [type], [type], [type] ];
+      return [
+        [type],
+        [type],
+        [type],
+        [type]
+      ];
     }
   }
 
   @override
   void rotate() {
-    if (this.rotation == Rotation.Normal || this.rotation == Rotation.SixOClock) {
+    if (this.rotation == Rotation.Normal ||
+        this.rotation == Rotation.SixOClock) {
       this.x++;
     } else {
       this.x--;
     }
     super.rotate();
   }
-
 }
