@@ -1,8 +1,6 @@
-
 import 'package:tetris/model/tetromino.dart';
 
 class TetrominoO extends Tetromino {
-
   @override
   TetrominoType get type => TetrominoType.o;
 
@@ -13,8 +11,10 @@ class TetrominoO extends Tetromino {
   int get width => 2;
 
   @override
-  List<List<TetrominoType>> get blocks {
-    return [ [type,type], [type,type] ];
+  List<List<TetrominoType?>> get blocks {
+    return [
+      [type, type],
+      [type, type]
+    ];
   }
-
 }
